requirements.txt
tkinter
tkinterdnd2
Pillow
